Change some tuples to pairs and typedef for readability
Change every get_min_dimensions to return "width_height", which is
hopefully obviously a pair containing a width and height.

Change tuples in the writes map to "address_auxpage", which is
maybe a more descriptive type of a pair containing an address and an
aux page.

Inspired by http://maintainablecode.logdown.com/posts/158531-stdpair-considered-harmful
    I kind of cheated - I used std::pairs but gave them descriptive typedefs.