fix crash if text is dragged from Chrome
If the drag-and-dropped filename has no leading "/", assume it is
drag-and-dropped text.  This matches the behavior of Chrome on Ubuntu,
for which dragging selected text onto GLFW is passed to the file drop
callback.

If the dragged selected text starts with "/", the drop callback will
still attempt to open it as a filename.